Start a new branch (from revision 1294) to fix various problems with the current term representation.  It's currently hard and messy to simplify formulas/bechmarks that involve bit-vectors and even more so to deal with benchmarks that require the egraph and bitvector solvers.
Here are some issues related to this problem:

On many benchmarks in the QF_AUFBV/brummayere series, the egraph
contains duplicate terms that the bit-vector solver knows are
equal. But the egraph does not know this (e.g., terms that are mapped
to identical bit-vector constants). This happens because the bv-solver
does some crude merging and boolean propagation of its own but the
results don't get propagated to the egraph or SAT solver.

Most SMT benchmarks in the QF_BV or QF_AUFBV categories use bit-vector
of size 1 as Booleans. Yices does not deal with them efficiently. The
context does not apply the usual boolean simplifications when the boolean
structure of a benchmark is hidden in bit-bitvectors.


Goals:
- use more uniform data structures to represent all the terms.
- remove the mapping from terms to variables in the term table.
- remove the separate representation of polynomials and power products
  (store power products directly in the term table).
- develop better simplification techniques for the context.



git-svn-id: svn+ssh://rfs.csl.sri.com/svn/private/yices2/branches/new-terms@1295 0c041c21-16df-db11-9388-003048836090