Fixed a bug in matrices.c: in matrxi_eliminate_fixed_variables, index j was declared unsigned (uin32_t) instead of signed (int32_t).
Fixed a bug in smt_core.c: we must reset theory_conflict (not just
the inconsistent flag) in internalization_start.

Added a new simplification step in simplex.c (and corresponding 
functions in poly_buffer): if a top-level equality (p == 0) is 
asserted, and p is of the form (b.x - c) == 0 then we can convert
p == 0 into two constraints on x: (x <= c/b) and (x >= c/b), which
is cheaper and enables early detection of conflict.



git-svn-id: svn+ssh://rfs.csl.sri.com/svn/private/yices2/branches/new-terms@1684 0c041c21-16df-db11-9388-003048836090