In yices_reval and context_solver: - forced all verbose-mode data to be printed on stderr rather than stdout. - disabled the 'ok' acknowledgment unless --verbose is given and   the solver is run without an input file (interactive mode).
Fixed support to tuple-select so that all (select tuple i)
requires 1 <= i <= number of components in tuple. Before this,
the API used 0 <= i < number of components, but that was not
consistent with the parser/term_stak.


Changed to format for function values in model_printer (in an attempt
to make the values easier to parse, for example, by SAL).





git-svn-id: svn+ssh://rfs.csl.sri.com/svn/private/yices2/branches/new-terms@2447 0c041c21-16df-db11-9388-003048836090