Removed smt2_lexer and smt2_parser from the build (and the relevant tests).
Fixed the source to prevent compilation warnings (when compiling on
juror30 using gcc-4.6.1). This version of GCC produces warnings for
variables assigned but not used. There are several places in Yices,
where variables are assigned and used only in 'assert', so we get
warnings when compiling in release mode. Fixed this by adding #ifndef
NDEBUG everywhere. This makes the code uglier.





git-svn-id: svn+ssh://rfs.csl.sri.com/svn/private/yices2/branches/new-terms@2484 0c041c21-16df-db11-9388-003048836090