Fixed a bug in egraph.c (cf. wd/bugs/segv1.ys, reported by Richard Uhler.
- the scenario was similar to 
    (push)
    (assert (= x y))
    (assert (... (f x x) ...)
     ...

    (pop)

  The assertion (= x y) causes the classes of x and y to 
  be merged.  Assuming y is the root term when then,
  when (f x x) is created, it's added to the congruence 
  table and to the parent vector for y's class.

- when (pop) is called, we must remove (f x x) from
  the congruence table. The bug was to undo the merge
  of x and y, before removing (f x x).  When we attempted
  to remove (f x x), the signature was wrong (i.e., it was
  still [f y y], so the term wasn't found in the congruence
  table and it was kept around also in y's parent vector).

- fix: when a composite term is created at decision_level>0,
  we add it to the undo stack. (Previously, I did this
  only when decision_level > base_level, but that was not
  enough).


Fixed a bug in internalization of tuple terms. When
instantiating the skolem axiom of a tuple t, we 
create variables x1 ,,, x_k in the Egraph and add 
the constraint (t = (tuple x1 ... x_k)). We must 
recursively add type constraints for x1 ... x_k.
The previous version did not recursively call 'add_type_constraints'.
It recursively called 'skolemize_tuple' for each x_i that has
a tuple type. That was not enough, we must add constraints also
if x_i has a finite type.


Changed the way constants of scalar (or uninterpreted) types
are displayed in (show-model). Example:

  (define-type S (scalar A B))
  (define x::S)
  (define y::S)
  (assert (/= x y))
  (check)
  (show-model)

should produce
  (= x A)
  (= y B)

or the other way around. This is better for input in the Yices 2
language. For SMT-LIB 1.2 input (e.g., in QF_UF) we print something
like

  (= x const!1)
  (= y const!2).


Added a minimal example to be sent to Karpenkov.


   



git-svn-id: svn+ssh://rfs.csl.sri.com/svn/private/yices2/branches/new-terms@2546 0c041c21-16df-db11-9388-003048836090