Fixed a bug in symbol_table:
In stbl_extend, we must copy the records in the right order. The
previous version was incorrect: because of the free list, we can't
just scan the record bank in order.


Added a simplification for mk_lambda:

  (lambda (x) (f x)) --> f 

provided x has the same type as domain of f.


Added support for pretty printing values and models.




git-svn-id: svn+ssh://rfs.csl.sri.com/svn/private/yices2/branches/new-terms@2557 0c041c21-16df-db11-9388-003048836090