Incorrect assert in balanced_arith_buffers.c (test for equality).
Fixed push/pop in simplex.c to make sure the simplex solver and the
eqprop manager are in sync: they have the same base_level and
decision_level.

Fixed a bug in egraph.c: in restore_eterms, there's a special
case that caused the atom removal to break and the assert(..)
to fail. For a predicate t that's asserted at the top level,
we do 

   egraph->terms.thvar[t] = const_bvar
   egraph->classes.thvar[c] = null_thvar

where c = class of t. For such a t, we then have

   egraph_term_thvar(egraph, t) = null_thvar
   egraph_term_base_thvar(egraph, t) = const_bvar


The code failed in this case, and we shouldn't try
to delete the atom attached to const_bvar.




git-svn-id: svn+ssh://rfs.csl.sri.com/svn/private/yices2/branches/new-terms@2881 0c041c21-16df-db11-9388-003048836090