The previous update added flattening for (eq t u) when t and u are arithmetic variables.  But flattening interferes with the simplification rules for if-then-else, since  these rules attempt to simplify constraints of the from  (eq (ite c a b) x).
Modified the flattening code to prevent flatteing of
(eq t u) if t or u is a an if-then-else term.

Other change in context.c to deal with if-then-else
occurring in ARITH_EQ0 atoms (i.e., (= (ite c a b) 0)).

Added a flag in test_smt_internalizer to enable/disable
pretty printing of the assertions.





git-svn-id: svn+ssh://rfs.csl.sri.com/svn/private/yices2/branches/new-terms@2896 0c041c21-16df-db11-9388-003048836090