Renamed the input files used by table_builder to xxx_input_tables.h instead of xxx_parser_tables.h (to reduce the risk of confusing them with the output form table_builder, which are named xxx_parse_tables.h).
Renamed smt2_ast to parenthesized_expr and added an smt2_expression module.

Added new options to smt2_commands so that main can set the initial
value of :output-success in the SMT2 interpreter.

Added a new type of atomic token to yices_pp so that we can print 
strings with quotes (as needed for SMT2).

Removed the unused fvar_collector structure from the term_substitution
data structure and module.



git-svn-id: svn+ssh://rfs.csl.sri.com/svn/private/yices2/branches/new-terms@2997 0c041c21-16df-db11-9388-003048836090