Changed model construction in egraph.c to use the new fresh_value_maker module.
Moved the 'add_type_constraint' code that used to be in context.c into
egraph.c. This is required because new terms may be created by
theory solvers in the egraph (without going through the internalization
procedure in context.c).



Improved the model construction to keep track of variables that
are relevant but are not internalized. Example:

  (define a::bool)
  (define b::bool)
  (assert (/= a b))
  (check)

The assertion is reduced to (b == (not a)): so there's nothing
generate in the smt_core or any solver. All we have in the
internalization table is 
   b --> (not a)
   a --> root

The previous model construction did not build a mapping for a,
in such cases. So (show-model) gave

   (= b true).

New version: if 'a' is not internalized, check whether 'a' is relevant
(i.e., present in the internalization table and if so, assign a
default value to a.

To avoid code duplication: moved the vtbl_make_object function to
concrete_values.c.




git-svn-id: svn+ssh://rfs.csl.sri.com/svn/private/yices2/branches/new-terms@3044 0c041c21-16df-db11-9388-003048836090