Increased the threshold to skip dsolver_is_feasible to 2000 (from 100).
Changed construction of eager lemmas to avoid running out of
memory. For a new atom (x >= a) or (x <= a), the previous version
created the lemmas involving this atom an all other atoms on x.
The new version creates at most two atoms. Among all the other
atoms on x, it searches for two atoms:
  atom1 of the form   (x >= a1) or (x <= a1)  with a1 <= a
  atom2 of the form   (x >= a2) or (x <= a2)  with a <= a2

and generates the binary atoms relating 'new-atom' and 'atom1',
and 'new-atom' and 'atom2'.