Merged the new API functions from the master branch (described below). Fixed merge conflicts in src/Makefile
Also fixed a bug in convert_value_array: uninitialized t if n = 0.

Commit message from the master branch
-------------------------------------
Added conversion of values to constant terms + implemented
the API functions that use it:

  term_t yices_get_value_as_term(model_t *mdl, term_t t);
  int32_t yices_term_array_value(model_t *mdl, uint32_t n, const term_t a[], term_t b[]);

Fixed a bug in model_eval: we don't want to create temporary values
in eval->vtbl, since we need the values after the evaluator is
deleted.

Fixed comment in the remove_trailing_whitespaces script.