Bug fix in literal_collector: some Boolean terms may be visited twice e.g. first as terms then as formulas. So we need to caches: one for terms and one for formulas. (This is safe, but possibly not optimal).
Added missing update of scan_idx in efsolver: this is intended to
ensure fairness in case there are serveral universal constraints.
(This change makes no differences on the benchmarks we have).

Added support in elim_susbt to eliminate propositional variables (by
treating atom P as (P := true) and (not P) as (P := false).