Fix in src/reader.c: use getc instead of getc_unlocked.
Updates to yices_smtcomp: use the official API to create
context and params array and to build the model. Still
some dependencies on internal stuff for statistics and
results.

Updates to yices_reval: reduced the use of direct calls
to functions in context.h etc.

Additions to yices_api.c: new functions to print the
value of a term and to get the push_level.

Removed unused headers from yices_smtcomp_mt.c.