Added a rewriting rule in ef_analyze to apply distributivity:
This rewrites
   (or a[0] ... a[n-1] (and b[0] ... b[m-1]))

to (and (or a[0] ... a[n-1] b[0])
         ...
        (or a[0] ... a[n-1] b[m-1])).

We use a heuristic to decide when to apply this rule: currently,
the rule applies if there's only one conjunct (and b[0] ... b[m-1])
in the (or ....).

Disabled test of implicant computation from efsolver.c + reduced
output (don't print the conditions on X).