Bug fix: in smt_core's theory_propagation, we must return false if  s->th_ctrl.propagate(s->th_solver)  returns false or if s->th_ctrl.propagate(s->th_solver) sets s->inconsistent to true.
This may happen if base_propagate is called with Egraph + bitvector
and an equality from the egraph is propagated to the bitvector
solver, and the bitvector solver detects that the equality is
false at the base level. In such a case, the bvsolver adds the
empty clause in the core s, which just sets s->inconsistent to true.
But the egraph's propagate returns true.