Changed smt2_term_stack once more to deal with names in declarations and definitions.
In (declare-sort <name> ...) and (define-sort <name> ...), we can't
tell whether <name> if for a simple sort or a sort constructor. So we
now complain if <name> is already used as a sort name or a sort-name
constructor.

In (declare-fun <name> ...) and (define-fun <name> ...), we have
the same problem. We can't tell whether <name> is intended as a
term name or a function name, so we disallow both.