Bug fix: in assert_arith_bineq, we try to simplify equalities of the form
   (= x (ite c a b))

to either (and c (= x a)) or (and (not c) (= x b)) when we know
that (= x b) or (= x a) is false.

In the first case, if the simplification succeeds, we assert c then
the equality (= x a). Asserting c may modify the substitution table so
we must apply substitution to 'x' and 'a' after processing 'c'.
(this subsitution was missing).