Bug fix: better processing of conditionals/removal of an incorrect assert.
In map_conditional_to_eterm and map_conditional_to_arith, the code
was checking whether one of the conditions c->pair[i].cond is true.
In such a case, we can assume that all the preceding conditions
 c->pair[0] ... c->pair[i-1] are false. However, this does not
necessarily mean that they all are internalized to false_literal.

An example is:

   (assert a)
   (assert (= x (if b 10 (if a 0 20))))

Assuming b and a are disjoint (i.e., (a and b) is false) then
the term (if b 10 (if a 0 20)) is converted to a conditional:

   cond b --> 10
        a --> 0
     else --> 20
   end

In this example, a is true but there's no reason to assume that
b is internalized to false_literal by the theory solver.