Bug fix in bit_term_conversion.c: invalid assert in convert_term_to_bit.
When term 't' is converted to a bit 'x', then function
convert_term_to_bit tries to also store that 'x' is mapped to 't'
in nodes->map. The bug was to assume that 'x' always has positive
polarity.

Other changes: the reverse conversion (i.e., convert_bit_to_term) is
now implemented in term_manager.c. This makes much more sense, as the
term_manager already provides constructors for binary XOR, binary OR,
and bitselect, which is what the bit-to-term conversion needs.