Bug fix: added a function to reset possible side effects of final_check.
The egraph's final check uses model reconciliation. It tries to merge
terms and equivalence classes so that the egraph agrees with a
satellite solver's model.

If this succeeds, new equations are present in the egraph's
propagation queue. These equations are due to the reconciliation
process. They must be kept in case we have to build a model.

However, they must be removed on the next call to push/pop/assert.
In most cases, the next call to push/pop/assert will call
egraph_backtrack to the reconciliation equalities get removed
automatically. But if the final_check succeeds at the base_level,
there's nowhere to backtrack, and the reconciliation equalities
remain present. This is incorrect and unsound.

To fix this, I've added a function 'clear' to the theory solver
interface. This function is called when 'smt_clear' is applied
to the core. The function 'clear' must undo any side effect
of the last successful call to 'final_check'.