Bug fix: cleanup ub/lb tags in simplex_pop.
In some circumstances, the simplex solver's simplex_pop function
left ub/lb tags set on a variable when the tags should have been
cleared. This results in errors in subsequent make_feasible (i.e.,
a variable is thought to be assigned to its upper or lower bound
when it is not).

This could happens in cases where we have

   (push)
   (assert)
   (push)
   (assert)
   (check-sat)
   (pop)
   (pop)

The first of the two (pop) must clean the lb/ub tags if it
rolls back the fix_ptr in the solver's bound stack (bstack), because
the second (pop) may remove bounds in stack[fix_ptr, top-1] but
it will assume that these bounds have not been processed and keep
the lb/ub tags of variables in these bounds unchanged.