Bug fix: take care of signs when building a power product in bvpoly_dag.
When building a dag_pprod for something like (bvmul (u^d) (v^e)), we can't
assume that the node occurrences for u and v are positive. We ensure
that all nodes inside the product dag have positive signs, and we
return +/- the product dag depending on exponents and signs of u and v.
For example, if d=2 and e=3, the result depends on whether u and v
are mapped to positve and negative nodes:

  u --> +nu   v --> +nv    product --> + [pprod +nu^2 +nv^3 ]
  u --> -nu   v --> +nv    product --> + [pprod +nu^2 +nv^3 ]
  u --> +nu   v --> -nv    product --> - [pprod +nu^2 +nv^3 ]
  u --> -nu   v --> -nv    prodcut --> - [pprod +nu^2 +nv^3 ]