Bug fix: in bvpoly_compiler.c, we can't assume that mk_bvmul and mk_bvadd always create fresh variables (because the nodes in the bvpoly_dag are not a normal form).
For example, it's possible to have nodes of the form

n4: [LEAF u!k]
n5: ...
n6: [PROD +n4 +n5]
n7: [MONO <constant> +n5]

then n6 and n7 will compile to the same variable if u!k
is equal to <constant>. Bug 0020 in the Mantis database
illustrates this case. It's now added as a regression
test.