Simplified/improved the literal_collector.
When X is a Boolean variable, we don't attempt to keep atoms of the
form (= X <term>) unless <term> is itself a simple Boolean term. Simple
terms include: true/false, Boolean variable, uninterpreted predicates,
and tuple/bit select.

In particular, we should never get terms of the form
      (= X <arithmetic constraint>)
   or (= X (or ....))
in an implicant.

We can't process such terms in the model-based projection module
because applying the substitution (X := <constraint>) may introduce
<constraints> that the arith_projector cannot handle.  For example if
we have (X := (eq t u)) and (not X) occurs somewhere, then the
substitution will introduce (not (eq t u)) and we can't handle this
literal.