Bug fix: incorrect simplification of n-ary XORs.
In function mk_xor, a term like (xor a1 ... a_n true) gets
simplified to (not (xor a1 ... a_n)). If n is equal to 2, we
want to treat this in the same way as in a call to

  mk_binary_xor(a1, a2)

which gets translated to opposite_term(mk_iff(a1, a2)). There
was one negation missing: the function was returning a term equal
to mk_iff(a1, a2).