Bug fixes in term-exploration API + more explanations in yices.h.
The Boolean constant false is internally encoded as (not true),
but we treat false as an atomic Boolean constant in the API.
Some functions in term_explorer.c did not do it properly.

Issues with index i in (select t i): internally, the index is
between 0 and n-1. In the API and language, i must be between 1
and n. To be consistent, yices_proj_index(u) must return a number
between 1 and n if u is a tuple projection.