Removed 'bool' from yices_types.h.
We can't use the type 'bool' in C unless we import <stdbool.h>, which
may not be supported by all compilers. Also, not sure that the C
header <stdbool.h> is compatible with C++ (since C++ has the type
bool).

Fixed search_parameters.c to handle boolean flags as integers.