Updates to src/Makefile.
On linux, compile with flag -U_FORTIFY_SOURCE (otherwise, the
runtime library will refer to __longjmp_chk and other variants
of standard functions and this may cause link-time error if
these special functions are not available).

Added flag --Wl,--no-as-needed when building with MODE=gperftools.
Otherwise, the -lprofiler library is not linked into the binaries
(on linux).

These flags are set to other values by default on Ubuntu.
This is explained at https://wiki.ubuntu.com/ToolChain/CompilerFlags.