Hack to handle division by non-constant in SMT2.
We should create a new term constructor for ordinary division.
For now, since we don't have the time, we convert division by
something that's not a nonzero rational into (div x y).
This is a hack: (div x y) has type Int and denotes integer
division. MCSAT must recognize this and handle this case
differently than integer division when x or y are not integer.

Also renamed regression test that timeout in nia.