More error checking for MCSat.
MCSat doesn't support incremental solving. We now check for this in
yices_smt2 and in smt2_commands.c.

Also improved sthe context_config code to check for logics that MCSat
supports and to report an error if MCSat and an incremental mode are
selected.

Updates to the doc to explain these changes.