Fix bug reported by Hoang Le.
Corner case in bvsolver.c: a combination of delayed processing
of variables, followed by pop, and simplifications that hide
variables.

1) Bitblasting works in several passes.

First, we mark all variables that need bitblasting:
- every variable that occurs in an atom
- variables x and y that are merged (i.e., x == y is a top-level
  constraint)
- variable x that's attached to an E-graph term.

Then we, compile polynomials to simpler terms (using bvcompiler).  It
turns out that some variables that are marked as useful in the first
phase may be simplified away by the compilation process (so they never
get bitblasted).

2) Some variables must be stored in solver->delayed_queue

If a variable x is created at base_level k but is processed at a
higher level than k, then we map an array of literals to x at
level k' > k. When we exit level k' (on 'pop'), we must delete
the literal array mapped to x because the literals for x are
removed from the SAT solver.  We keep track of x in the
solver->delayed_queue.

The bug was to store 'x' in the delayed_queue only on a call
to bitblast variable. This is not enough: x may have a literal
array attached to it (because we marked x as a useful variable
in phase 1 of bitblasting) but 'bitblast_variable' may never
be called for 'x' (because the bvcompiler removed x).