Cleanup
Updated headers to the GPL license in some source files that
still had the old license.

In the smt2 frontend, the global variables ctx_parameters and parameters
are now part of __smt2_globals (to be consistent).

Moved the 'ctx_param_t' and relatives out of frontend/common
and into a separate context/context_parameters module.