Improvements to the context simplifications.
1) detect when a disjunct is trivially true

When processing (or a_1 ... a_n), we check whether one of a_i
is known to be true. In such a case, the disjunct is true and
we skip internalization of the other a_j's.

2) better estimate of path lengths for difference-logic

The upper bound on the max length of a simple path is less
crude. So we should be able to use the Floyd-Warshall IDL
solver on more instances.

3) Other small fixes: typo in Makefile.