Handle a special case in check_assumptions:
If a contradiction is detected in yices_assert_formula, before the smt
solver is called, then the call to derive_unsat_core does not produce
anything.

Fixed this by checking whether an assumption[i] is false, which seems to
be the only case where assert_formula(indicator[i]) can detect unsat.
Just to be safe, we check for it and generate a safe but imprecise
core if assumption[i] is not false.