Started a hash map with support for push/pop.
Function bv_compiler_remove_vars is not sufficient to undo the effects
of a round of compilation because compilation does not always create
new variables. If the compilation maps a variable x to some other
variable y, then the mapping [x --> y] must be undone on the next call
to smt_pop (even if x and y are kept after smt_pop).