Correction to clean interrupt
In clean interrupt mode if the query is UNSAT even before calling yicec_check_context, an extra call to smt_pop can happen (due to context_clear_unsat). Since this pop does not have a corresponding push (since start_search hasn't been called), it leads to a segfault in such a case.

The fix is simple. Use a boolean interrupt_push to decide whether to pop or not.