Bug fix: extend partial assignments from cadical to full assignments.
The bit-vector solver doesn't assign a value to a variable
x unless all the bits of x are set to either true or false by
the backend solver. If one bit of x is 'unknown' the bit-vector
solver sets the value of x to 'unknown', which may lead to an
incorrect model later on.