Adds --abort-on-int cmdline flag to yices executable.
We use the yices executable (and not the yices_smt2 executable) as a
sub-process solver from our primary application.  Our application
expects normal cleanup to the foreground process group to occur on
Ctrl-C / SIGINT, but currently it will exit and leave the yices
process running in the background because yices does not exit on this
event.

This small patch adds the cmdline flag that we can use from our
application to run yices in a mode that will exit on these events.
The default is the current behavior so this change should be fully
backward-compatible.